mon: replace ConnectionTracker with new implementation
