+++ src/mon/Elector.cc
@@ -60,7 +60,7 @@ static ostream& _prefix(std::ostream *_dout, Monitor *mon, epoch_t epoch) {
 
 Elector::Elector(Monitor *m) : logic(this, ElectionLogic::CLASSIC,
 				     &peer_tracker, m->cct),
-			       peer_tracker(this, 12*60*60), // TODO: make configurable
+			       peer_tracker(this, m->rank, 12*60*60), // TODO: make configurable
 			       ping_timeout(2),
 			       mon(m), elector(this) {
   bufferlist bl;
